refactor: extract milestone-ids and guided-flow-queue from guided-flow.ts
- Extract milestone ID utilities (MILESTONE_ID_RE, generateMilestoneSuffix,
  nextMilestoneId, extractMilestoneSeq, parseMilestoneId, milestoneIdSort,
  maxMilestoneNum, findMilestoneIds) into milestone-ids.ts (~95 lines)
- Extract queue management (showQueue, handleQueueReorder, showQueueAdd,
  buildExistingMilestonesContext) into guided-flow-queue.ts (~445 lines)
- Add re-exports from guided-flow.ts to preserve public API
- Fix circular dependency: queue-order.ts now imports milestoneIdSort
  from milestone-ids.js instead of guided-flow.js
- guided-flow.ts reduced from 1611 to 1144 lines
